[tests] Re-added debug mode in tests
The EXTRA_FLAGS needs some clean before running tests.
This commit allows to compile twice for this situation, but it is not perfect
because EXTRA_CFLAGS can still cause some problems. The tests Makefile should
be able to detect this and to group builds and tests with each own EXTRA flags.