Remove disconnect memory leak
When disconnecting, the client needs to be disabled and assuming it is a client
created by the broker, it needs to be deleted. Previously the client was
directly removed from the broker, thereby losing any reference to it and
it could never be deleted.

This changes that behavior to purely invalidate the client, and the next time
through the broker loop it will be properly cleaned up. This also fixes the
loop code to properly erase the client from the list, and not break out early
the first time it finds a disconnected client.

This doesn't fix everything, in that it seems like there's another bug that it
is deleting the client without a flag check. However I'm leaving that for a
separate fix.

I have tested this on my ESP32 by uploading this version of the code, running
a client in an infinite loop which connects, publishes, disconnects. With the
old code I can steadily see memory usage linearly leaking over time. With the
new version it stays stable after the first couple times.

More in the linked issue.

Fixes #73